Move graph related classes to own file
